Add files via upload
UML Diagram representing Builder design pattern.
In this example I demonstrated TPH(Table-per-Hierarchy) object creation replicating following scenario. 
We have 2 base models that later on can be mapped to tables: 
1) SystemUser
2) Doctor
Relationship between these two tables is one to one. Doctor table is used to handle additional data that is not related to SystemUser table. 
SystemDoctor class that inherits from SystemUser is used to store all properties that are associated with SystemUser and Doctor class. In order to save data to underlying database we have to create 2 separate objects( SystemUser and Doctor).